made all mutations require authentication from user

+++ src/resolvers/Mutation.js
@@ -45,20 +45,14 @@ const Mutation = {
             token: jwt.sign({ userId: user.id }, 'thisisasecret')
         }
     },
-    async deleteUser(parent, args, { prisma }, info) {
-       const userExists = await prisma.exists.User({ id: args.id })
+    async deleteUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
 
-         if (!userExists) {
-             throw new Error('User not found')
-         }
-
-         return prisma.mutation.deleteUser({ 
-             where: {
-                 id: args.id 
-                }
-            }, info)
-
-
+        return prisma.mutation.deleteUser({ 
+            where: {
+                id: userId 
+            }
+        }, info)
     },
     async updateUser(parent, args, { prisma, request }, info) {
         const userId = getUserId(request)
@@ -87,7 +81,20 @@ const Mutation = {
     }, info)
 
     },
-    async deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        // check if post exists using token as deletepost doesnt take userid as an arg
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!postExists) {
+            throw new Error('Unable to delete post')
+        }
+
         return prisma.mutation.deletePost({ 
             where: {
                 id: args.id 
@@ -95,7 +102,19 @@ const Mutation = {
            }, info)
 
     },
-    async updatePost(parent, args, { prisma }, info){
+    async updatePost(parent, args, { prisma, request }, info){
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!postExists) {
+            throw new Error('Unable to update post')
+        }
+
         return prisma.mutation.updatePost({ 
             where: {
                 id: args.id
@@ -103,7 +122,8 @@ const Mutation = {
             data: args.data,
             }, info)
     },
-    async createComment(parent, args, { prisma }, info) {
+    async createComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
         return prisma.mutation.createComment({
             data: {
                 text: args.data.text,
@@ -114,20 +134,42 @@ const Mutation = {
                 },
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 }
             }  
          }, info)
     },
-    async deleteComment(parent, args, { prisma }, info) {
+    async deleteComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!commentExists) {
+            throw new Error('Cannot delete comment')
+        }
         return prisma.mutation.deleteComment({ 
             where: {
                 id: args.id
             }
          }, info)
     },
-    async updateComment(parent, args, { prisma }, info) {
+    async updateComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!commentExists) {
+            throw new Error('Cannot update comment')
+        }
         return prisma.mutation.updateComment({
             where: {
                 id: args.id
